move fn to separate resolver
